Interfacing Temperature Sensor Using Arduino.
The project is a temperature sensor made with an Arduino UNO and LM35 sensor and few other 
components. The main aim of the project is to detect the temperature of the surroundings and display it on the LCD.The working of the project starts with the LM35 sensor that senses the change in temperature of the surrounding, and uses that temperature difference to produce a voltage signal which is processed by the Arduino to give a digital output displaying the temperature of the given surrounding.