Interfacing Temprature Sensor Using Arduino
The project is a temperature sensor made with anArduino UNO and LM35 sensor and few othercomponents. The main aim of the project is to detectthe temperature of the surroundings and display it onthe LCD.The working of the project starts with the LM35sensor that senses the change in temperature of the surrounding, and uses that temperature difference toproduce a voltage signal which is processed by theArduino to give a digital output displaying thetemperature of the given surrounding.